fix(sandbox): harden dcode probe diagnostics
