+++ src/lib/actions/sandbox/connect-flow.test.ts
@@ -8,10 +8,6 @@ import {
   createConnectHarness,
   requireDist,
 } from "../../../../test/support/connect-flow-test-harness";
-import {
-  DCODE_MANAGED_EXEC_LAUNCHER,
-  DCODE_MANAGED_EXEC_MISSING_DETAIL,
-} from "./connect-inference-route-probe";
 
 describe("connectSandbox flow", () => {
   let exitSpy: MockInstance;
@@ -439,35 +435,6 @@ describe("connectSandbox flow", () => {
     expect(errorOutput).toContain("did not return a trusted result");
     expect(errorOutput).toContain("OpenShell CA bundle missing or unreadable");
     expect(errorOutput).not.toContain("after DNS and route repair");
-    expect(exitSpy).toHaveBeenCalledWith(1);
-  });
-
-  it("fails closed with rebuild guidance when a DCode sandbox lacks the trusted helper (#6192)", async () => {
-    const harness = createConnectHarness({
-      agentName: "langchain-deepagents-code",
-      registryEntry: {
-        provider: "nvidia-prod",
-        model: "nvidia/nemotron-3-super-120b-a12b",
-      },
-      inferenceGetOutput:
-        "Gateway inference:\n  Provider: nvidia-prod\n  Model: nvidia/nemotron-3-super-120b-a12b\n",
-      inferenceProbeResponses: [`exec: ${DCODE_MANAGED_EXEC_LAUNCHER}: not found`],
-      sessionAgent: { name: "langchain-deepagents-code" },
-    });
-
-    await expect(harness.connectSandbox("alpha")).rejects.toThrow("process.exit(1)");
-
-    expect(harness.applyVmDnsMonkeypatchSpy).not.toHaveBeenCalled();
-    expect(harness.runSetupDnsProxySpy).not.toHaveBeenCalled();
-    expect(harness.runOpenshellSpy).not.toHaveBeenCalled();
-    expect(harness.spawnSyncSpy).not.toHaveBeenCalledWith(
-      "openshell",
-      ["sandbox", "connect", "alpha"],
-      expect.any(Object),
-    );
-    const errorOutput = harness.errorSpy.mock.calls.flat().join("\n");
-    expect(errorOutput).toContain(DCODE_MANAGED_EXEC_MISSING_DETAIL);
-    expect(errorOutput).toContain("sandbox inference route is not known healthy");
     expect(exitSpy).toHaveBeenCalledWith(1);
   });
